// types/Job.ts
export interface IJob {
  createdBy: string
  createdDate: string
  lastModifiedBy: string
  lastModifiedDate: string
  id: 1
  departmentName: string
  jobTitle: string
  jobDescription: string
  salary: 15.5
  position: string
  requirements: string
  postedDate: string
  closedDate: string
  location: string
  exp: 0
  active: true
}